# server/openapi.yaml
openapi: 3.0.0
info:
  title: Communicator API 2026
  description: >
    API REST + WebSocket para sistema de mensajería instantánea y
    llamadas peer-to-peer (WebRTC). Soporta chat 1-a-1, grupos,
    presencia en línea, y señalización de llamadas de voz/video.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor Local de Desarrollo (ADB reverse)
  - url: http://192.168.XX.XX:3000
    description: Red WiFi local (reemplazar XX.XX con IP del servidor)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en /api/auth/login o /api/auth/register
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Credenciales inválidas
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              path:
                type: string
              location:
                type: string
...
    User:
      type: object
      required: [id, username, email, avatar_url, is_online, last_seen, tenant_id, active_workspace_id, is_suspended]
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
          example: juan_perez
        email:
          type: string
          format: email
          example: juan@example.com
        avatar_url:
          type: string
          nullable: true
          example: null
        is_online:
          type: boolean
          example: true
        last_seen:
          type: string
          format: date-time
          nullable: true
        tenant_id:
          type: integer
          nullable: true
          description: ID del tenant al que pertenece (alias de compat; hoy = workspace_id)
          example: 1
        active_workspace_id:
          type: integer
          nullable: true
          description: Workspace activo del usuario (seleccionado)
          example: 2
        workspace_ids:
          type: array
          nullable: true
          description: IDs de todos los workspaces a los que pertenece el usuario
          items:
            type: integer
          example: [2, 5]
        is_suspended:
          type: boolean
          description: Si la cuenta está suspendida (login bloqueado)
          example: false
...
    UserBrief:
      type: object
      required: [id, username, email, avatar_url, tenant_id, active_workspace_id]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        avatar_url:
          type: string
          nullable: true
        tenant_id:
          type: integer
          nullable: true
          description: ID del tenant al que pertenece (alias de compat; hoy = workspace_id)
          example: 1
        active_workspace_id:
          type: integer
          nullable: true
          description: Workspace activo del usuario (seleccionado)
          example: 2
        workspace_ids:
          type: array
          nullable: true
          description: IDs de todos los workspaces a los que pertenece el usuario
          items:
            type: integer
          example: [2, 5]
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
          description: Solo letras, números y guión bajo
          example: juan_perez
        email:
          type: string
          format: email
          example: juan@example.com
        password:
          type: string
          minLength: 6
          description: Mínimo 6 caracteres
          example: MiClaveSegura2026
        invitationCode:
          type: string
          minLength: 4
          maxLength: 64
          description: |
            Código de invitación opcional. Si se incluye, el usuario también se une
            al workspace del código como admin (además de su workspace personal).
            Formato típico: SLUG-RANDOM (ej: ACME-7XK2P4B9).
          example: ZAPATADEV-A1B2C3D4
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: juan@example.com
        password:
          type: string
          example: MiClaveSegura2026
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT con expiración de 7 días
          example: eyJhbGciOiJIUzI1NiIs...
        user:
          $ref: '#/components/schemas/UserBrief'
...
  responses:
    Unauthorized:
      description: Token JWT requerido o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/auth/register:
    post:
      summary: Registrar nuevo usuario
      description: |
        Crea un usuario con username, email y password.
        - Username: 3-50 caracteres alfanuméricos + guión bajo
        - Email: formato email válido
        - Password: mínimo 6 caracteres
        - Email y username deben ser únicos
      tags: [Autenticación]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Error de validación en campos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: El email o username ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: El email o usuario ya existe
        '429':
          description: Rate limit excedido (20 solicitudes / 15 min)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      summary: Iniciar sesión
      description: |
        Autentica al usuario con email y password.
        Devuelve un JWT con expiración de 7 días.
      tags: [Autenticación]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit excedido
  /api/auth/me:
    get:
      summary: Obtener perfil del usuario autenticado
      tags: [Autenticación]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ══════════════════════════════════════════════
  # MÓDULO DE UPLOAD DE IMÁGENES
  # ══════════════════════════════════════════════
...